+++ service/src/main/java/com/epam/esm/service/impl/TagServiceImpl.java
@@ -1,10 +1,10 @@
 package com.epam.esm.service.impl;
 
-import com.epam.esm.domain.GiftCertificate;
-import com.epam.esm.domain.Tag;
-import com.epam.esm.domain.dto.GiftCertificateDto;
-import com.epam.esm.domain.dto.TagDto;
-import com.epam.esm.domain.dto.serialization.DtoSerializer;
+import com.epam.esm.model.GiftCertificate;
+import com.epam.esm.model.Tag;
+import com.epam.esm.model.dto.GiftCertificateDto;
+import com.epam.esm.model.dto.TagDto;
+import com.epam.esm.model.dto.serialization.DtoSerializer;
 import com.epam.esm.exception.ErrorCode;
 import com.epam.esm.exception.ServiceException;
 import com.epam.esm.persistence.repository.TagRepository;
@@ -72,7 +72,7 @@ public class TagServiceImpl implements TagService {
         List<GiftCertificate> associatedGiftCertificates = tagRepository.findAssociatedGiftCertificates(tag.getId());
         page = PaginationUtil.correctPage(page, size, associatedGiftCertificates::size);
         return associatedGiftCertificates.stream()
-                .skip((page - 1) * size)
+                .skip((long) (page - 1) * size)
                 .limit(size)
                 .map(certificateDtoSerializer::dtoFromEntity)
                 .collect(Collectors.toList());
